Refactor fn parsing and support per-fn context
Any marshalled spooky function argument can now be either a bare
function or tuple: [Object, Function]. The members of the Object will
appear in the scope of the Function. So,
  [{ foo: 'bar', fiz: 'biz' }, function () { return foo; }]
translates to something like
  (function () {
    var foo = 'bar';
    var fiz = 'biz';
    return function () { return foo; };
  }())
in the CasperJS environment.

Refactor function parsing to use Function constructor only for checking
that the passed source can be eval'd as a single function. This allows
slightly better scope protection.