Marshal casper events across to spooky
Spooky now acts as a JSON-RPC server, line-buffering stdout of any
connected CasperJS instances. Any lines that are valid JSON-RPC 2.0
packets are handled by Spooky and other lines are written to Spooky's
stdout.

At present this mechanism is used to emit CasperJS events on the Spooky
instance. Call the on method of the Spooky instance to subscribe.