Implement stdio transport (fix #9)
By default, Spooky will now use the child's stdin/out as transport for RPCs. Set
options.child.transport to 'http' to use HTTP transport instead. HTTP transport
is probably going to be deprecated, however, as it was always a workaround for
stdin being buffered.

Specify the transport when running the tests with the TEST_TRANSPORT variable.
The default is 'stdio'.