Don't try to poll stdio again on `run.complete` (fixes #28)
In a misguided effort to allow a single spooky bootstrap instance service
multiple sessions, we reentered the polling loop in the stdio server.
Unfortunately, we did this on the `run.complete` event, which can fire multiple
times, if a script calls `start` and then `run` more than once. When the first
of these fired, the stdio server set a timeout to reenter the loop, which
promptly blocked waiting for stdin data that may never come.

We now do not attempt to poll stdin again after the initial run.